Filter out emoji data not needed to be transmitted
Prior to this change, I was using the whole emoji reference set when
there were duplicates and a bunch of info that was not needed

This change filters out the keys not needed and get the data into a more
useable format. It also introduces the tool that allows this process to
be repeated.